requirements.txt
# Core data/IO
pandas>=2.2
pyarrow>=16.0
numpy>=1.26

# HTTP + env + retries
httpx>=0.27
python-dotenv>=1.0
tenacity>=8.2

# ML/beregning
scikit-learn>=1.4
#xgboost>=2.0
#lightgbm>=4.3
#statsmodels>=0.14

# Plot (valgfrit – nice til EDA)
matplotlib>=3.8
plotly>=5.22
